// src/utils/type.ts
export type FacebookPageTokenUpdate = {
    client_id: string;
    client_secret: string;
    access_token: string;
    account_facebook: string;
    facebook_id: string;
}
// ...
export type AccountFacebook = {
    id?: string;
    gmail: string;
    facebook_id: string;
    name: string;
    active: boolean;
    have_app: boolean;
};

export interface AccountLinkedin {
    id?: string;
    linkedin_id: string;
    name: string;
    linkedin_profile_urn?: string;
    client_id: string;
    client_secret: string;
    redirect_uri?: string;
    access_token?: string;
    expires_at?: Date | null;
    gmail: string;
    active: boolean;
}

export type AppFacebook = {
    id?: string;
    account_facebook: string;
    app_name: string;
    client_id: string;
    client_secret: string;
    active: boolean;
    account?: AccountFacebook
};